moves get_notes to utils
